requirements.txt
alembic==1.13.1
Flask==3.0.3
Flask_Login==0.6.3
Flask_Migrate==4.0.7
Flask-JWT-Extended
pandas
Flask-CORS
python-dotenv
Requests==2.32.3
SQLAlchemy==2.0.30
Werkzeug==3.0.3
yfinance==0.2.40
supabase
numpy
google-cloud-profiler
gunicorn
python-dateutil
pykrx
google-cloud-storage
Flask-WTF